Just fix some minor build warnings
git-svn-id: http://svn.osgeo.org/mapguide/trunk@4737 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/PlatformBase/MapLayer/SelectionBase.cpp
@@ -679,7 +679,7 @@ MgStringCollection* MgSelectionBase::GenerateFilters(MgLayerBase* layer,
                 filterReserved = true;
             }
 
-            if ((selectionSize > 0) && ((filter.length() + selText.length() + OR_TEXT_SIZE) >= selectionSize))
+            if ((selectionSize > 0) && (((INT32)filter.length() + (INT32)selText.length() + OR_TEXT_SIZE) >= selectionSize))
             {
                 filters->Add(filter);
                 filter.clear();
